Rename to BuildDrop tasks.
