+++ src/Agent.Plugins/BuildDrop/BuildDropClient.cs
@@ -1,3 +1,10 @@
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.TeamFoundation.Build.WebApi;
 using Microsoft.TeamFoundation.DistributedTask.WebApi;
 using Microsoft.VisualStudio.Services.BlobStore.Common;
@@ -6,16 +13,10 @@ using Microsoft.VisualStudio.Services.BlobStore.WebApi;
 using Microsoft.VisualStudio.Services.Common;
 using Microsoft.VisualStudio.Services.WebApi;
 using Microsoft.VisualStudio.Services.Agent.Util;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Agent.Sdk;
 using Newtonsoft.Json;
+using Agent.Sdk;
 
-namespace Agent.Plugins.Drop
+namespace Agent.Plugins.BuildDrop
 {    
     // A wrapper of BuildDropManager, providing basic functionalities such as uploading and downloading drop artifacts.
     public class BuildDropClient
